Move `useBatchSpecCode`
