+++ client/web/src/enterprise/batches/batch-spec/edit/useBatchSpecCode.ts
@@ -8,9 +8,8 @@ import { debounce } from 'lodash'
 import { BatchSpec } from '@sourcegraph/shared/src/schema/batch_spec.schema'
 import { useDebounce } from '@sourcegraph/wildcard'
 
-import batchSpecSchemaJSON from '../../../../../../schema/batch_spec.schema.json'
-
-import { excludeRepo as excludeRepoFromYaml, hasOnOrImportChangesetsStatement } from './yaml-util'
+import batchSpecSchemaJSON from '../../../../../../../schema/batch_spec.schema.json'
+import { excludeRepo as excludeRepoFromYaml, hasOnOrImportChangesetsStatement } from '../../create/yaml-util'
 
 const ajv = new AJV()
 addFormats(ajv)
