+++ client/web/src/enterprise/batches/create/CreateOrEditBatchChangePage.tsx
@@ -59,6 +59,7 @@ import { EditorFeedbackPanel } from '../batch-spec/edit/editor/EditorFeedbackPan
 import { MonacoBatchSpecEditor } from '../batch-spec/edit/editor/MonacoBatchSpecEditor'
 import { LibraryPane } from '../batch-spec/edit/library/LibraryPane'
 import { ExecutionOptions, RunBatchSpecButton } from '../batch-spec/edit/RunBatchSpecButton'
+import { useBatchSpecCode } from '../batch-spec/edit/useBatchSpecCode'
 import { BatchSpecDownloadLink } from '../BatchSpec'
 
 import { GET_BATCH_CHANGE_TO_EDIT, CREATE_EMPTY_BATCH_CHANGE, CREATE_BATCH_SPEC_FROM_RAW } from './backend'
@@ -66,7 +67,6 @@ import { CodeInsightsBatchesIcon } from './CodeInsightsBatchesIcon'
 import { DownloadSpecModal } from './DownloadSpecModal'
 import { getTemplateRenderer } from './go-checker-templates'
 import { NamespaceSelector } from './NamespaceSelector'
-import { useBatchSpecCode } from './useBatchSpecCode'
 import { useExecuteBatchSpec } from './useExecuteBatchSpec'
 import { useInitialBatchSpec } from './useInitialBatchSpec'
 import { useNamespaces } from './useNamespaces'
